Create a network_analysis script that contains all the functions use to diagnose the netowrk and import them back into the simple_network script
